avoid using global singleton
