+++ src/generators/mediator.js
@@ -1,6 +1,7 @@
 // @flow
 
 import { activeElement } from '../dom'
+import { assert } from '../utils'
 
 export interface Mediator {
   currentName: string;
@@ -10,6 +11,10 @@ export interface Mediator {
 }
 
 export function generateMediator (Vue: any): Mediator {
+  const state = {
+    portal: null
+  }
+
   return new Vue({
     data: {
       stack: []
@@ -51,6 +56,16 @@ export function generateMediator (Vue: any): Mediator {
       replace (name: string): void {
         this.pop()
         this.push(name)
+      },
+
+      _setPortal(portal: any): void {
+        assert(!state.portal, '<modal-portal> is already created.')
+        state.portal = portal
+      },
+
+      _getPortal(): any {
+        assert(state.portal, 'You need to put <modal-portal> or set `autoMountPortal: true` plugin option.')
+        return state.portal
       }
     }
   })
